Restructure imports and package structure
